+++ src/common/constants.ts
@@ -58,3 +58,15 @@ export const is24HourWindowOver = (time: any) => {
 export const CONNECTION_RECONNECT_ATTEMPTS = 5;
 
 export const MEDIA_MESSAGE_TYPES = ['IMAGE', 'AUDIO', 'VIDEO', 'DOCUMENT', 'STICKER'];
+
+export const getObject = (arr: any, data: any) => {
+  const result: any = [];
+  if (arr && data) {
+    arr.forEach((obj: any) => {
+      data.forEach((ID: any) => {
+        if (obj.id === ID) result.push(obj);
+      });
+    });
+  }
+  return result;
+};
